+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.16.1] - 2019-05-25
+
 ### Fixed
 
 - Improve performance by removing dependencies and using function components.